backend/requirements.txt
# Flask 백엔드 의존성
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# 브라우저 자동화
playwright==1.40.0

# 스케줄링
APScheduler==3.10.4

# 텔레그램
python-telegram-bot==20.7

# 데이터베이스
SQLAlchemy==2.0.23

# 유틸리티
requests==2.31.0
python-dateutil==2.8.2
loguru==0.7.2
bcrypt==4.1.2
Flask-Limiter==3.5.0

# OCR (CAPTCHA 해결용)
easyocr==1.7.0
paddleocr==2.7.0
paddlepaddle>=3.0.0  # Python 3.13 호환
Pillow>=10.4.0  # Python 3.13 호환

# 개발 도구
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
